+++ th2_data_services/data_source/lwdp/commands/http.py
@@ -1090,6 +1090,7 @@ class DownloadMessagesByPageGzip(IHTTPCommand):
         sort: bool = None,
         response_formats: Union[List[str], str] = None,
         keep_open: bool = None,
+        streams: List[str] = None,
         # Non-data source args.
         max_url_length: int = 2048,
     ):
@@ -1102,6 +1103,7 @@ class DownloadMessagesByPageGzip(IHTTPCommand):
             sort: Enables message sorting within a group. It is not sorted between groups.
             response_formats: The format of the response
             keep_open: If true, keeps pulling for new message until don't have one outside the requested range.
+            streams: List of streams to search messages from.
             max_url_length: API request url max length.
         """
         response_formats = _get_response_format(response_formats)
@@ -1112,7 +1114,9 @@ class DownloadMessagesByPageGzip(IHTTPCommand):
         self._sort = sort
         self._response_formats = response_formats
         self._keep_open = keep_open
+        self._streams = streams
         self._max_url_length = max_url_length
+        _check_list_or_tuple(self._streams, var_name="streams")
 
     def handle(self, data_source: DataSource):
         page = _get_page_object(self._book_id, self._page, data_source)
@@ -1136,6 +1140,7 @@ class DownloadMessagesByPageGzip(IHTTPCommand):
             filename=self._filename,
             page=page,
             groups=groups,
+            streams=self._streams,
             book_id=self._book_id,
             sort=self._sort,
             response_formats=self._response_formats,
@@ -1170,6 +1175,7 @@ class DownloadMessagesByPageByGroupsGzip(IHTTPCommand):
         sort: bool = None,
         response_formats: Union[List[str], str] = None,
         keep_open: bool = None,
+        streams: List[str] = None,
         # Non-data source args.
         max_url_length: int = 2048,
     ):
@@ -1183,6 +1189,7 @@ class DownloadMessagesByPageByGroupsGzip(IHTTPCommand):
             sort: Enables message sorting within a group. It is not sorted between groups.
             response_formats: The format of the response
             keep_open: If true, keeps pulling for new message until don't have one outside the requested range.
+            streams: List of streams to search messages from.
             max_url_length: API request url max length.
         """
         response_formats = _get_response_format(response_formats)
@@ -1191,12 +1198,14 @@ class DownloadMessagesByPageByGroupsGzip(IHTTPCommand):
         self._page = page
         self._book_id = book_id
         self._groups = groups
+        self._streams = streams
         self._sort = sort
         self._response_formats = response_formats
         self._keep_open = keep_open
         self._max_url_length = max_url_length
 
         _check_list_or_tuple(self._groups, var_name="groups")
+        _check_list_or_tuple(self._streams, var_name="streams")
 
     def handle(self, data_source: DataSource):
         page = _get_page_object(self._book_id, self._page, data_source)
@@ -1213,6 +1222,7 @@ class DownloadMessagesByPageByGroupsGzip(IHTTPCommand):
             end_timestamp=self._end_timestamp,
             book_id=self._book_id,
             groups=self._groups,
+            streams=self._streams,
             sort=self._sort,
             response_formats=self._response_formats,
             keep_open=self._keep_open,
@@ -1257,6 +1267,7 @@ class DownloadMessagesByBookByGroupsGzip(IHTTPCommand):
         sort: bool = None,
         response_formats: Union[List[str], str] = None,
         keep_open: bool = None,
+        streams: List[str] = None,
         # Non-data source args.
         max_url_length: int = 2048,
     ):
@@ -1274,6 +1285,7 @@ class DownloadMessagesByBookByGroupsGzip(IHTTPCommand):
                   It's possible to add it to the CradleAPI by request to dev team.)
             response_formats: The format of the response
             keep_open: If true, keeps pulling for new message until don't have one outside the requested range.
+            streams: List of streams to search messages from.
             max_url_length: API request url max length.
         """
         response_formats = _get_response_format(response_formats)
@@ -1284,6 +1296,7 @@ class DownloadMessagesByBookByGroupsGzip(IHTTPCommand):
         self._start_timestamp = DatetimeConverter.to_nanoseconds(start_timestamp)
         self._end_timestamp = DatetimeConverter.to_nanoseconds(end_timestamp)
         self._groups = groups
+        self._streams = streams
         self._sort = sort
         self._response_formats = response_formats
         self._keep_open = keep_open
@@ -1291,6 +1304,8 @@ class DownloadMessagesByBookByGroupsGzip(IHTTPCommand):
         self._max_url_length = max_url_length
 
         _check_list_or_tuple(self._groups, var_name="groups")
+        _check_list_or_tuple(self._streams, var_name="streams")
+
 
     def handle(self, data_source: DataSource):
         api = data_source.source_api
@@ -1299,6 +1314,7 @@ class DownloadMessagesByBookByGroupsGzip(IHTTPCommand):
             end_timestamp=self._end_timestamp,
             book_id=self._book_id,
             groups=self._groups,
+            sterams=self._streams,
             sort=self._sort,
             response_formats=self._response_formats,
             keep_open=self._keep_open,
@@ -1334,6 +1350,7 @@ class GetMessagesByBookByGroups(SSEHandlerClassBase):
         sort: bool = None,
         response_formats: Union[List[str], str] = None,
         keep_open: bool = None,
+        streams: List[str] = None,
         # Non-data source args.
         max_url_length: int = 2048,
         char_enc: str = "utf-8",
@@ -1354,6 +1371,7 @@ class GetMessagesByBookByGroups(SSEHandlerClassBase):
                   It's possible to add it to the CradleAPI by request to dev team.)
             response_formats: The format of the response
             keep_open: If true, keeps pulling for new message until don't have one outside the requested range.
+            streams: List of streams to search messages from.
             char_enc: Encoding for the byte stream.
             decode_error_handler: Registered decode error handler.
             cache: If True, all requested data from lw-data-provider will be saved to cache.
@@ -1377,6 +1395,7 @@ class GetMessagesByBookByGroups(SSEHandlerClassBase):
         self._start_timestamp = DatetimeConverter.to_nanoseconds(start_timestamp)
         self._end_timestamp = DatetimeConverter.to_nanoseconds(end_timestamp)
         self._groups = groups
+        self._streams = streams
         self._sort = sort
         self._response_formats = response_formats
         self._keep_open = keep_open
@@ -1384,6 +1403,7 @@ class GetMessagesByBookByGroups(SSEHandlerClassBase):
         self._max_url_length = max_url_length
 
         _check_list_or_tuple(self._groups, var_name="groups")
+        _check_list_or_tuple(self._streams, var_name="streams")
 
     def _get_urls(self, data_source: DataSource):
         api = data_source.source_api
@@ -1391,6 +1411,7 @@ class GetMessagesByBookByGroups(SSEHandlerClassBase):
             start_timestamp=self._start_timestamp,
             end_timestamp=self._end_timestamp,
             groups=self._groups,
+            streams=self._streams,
             response_formats=self._response_formats,
             keep_open=self._keep_open,
             sort=self._sort,
@@ -1582,6 +1603,7 @@ class GetMessagesByPageByGroups(SSEHandlerClassBase):
         sort: bool = None,
         response_formats: Union[List[str], str] = None,
         keep_open: bool = None,
+        streams: List[str] = None,
         # Non-data source args.
         max_url_length: int = 2048,
         char_enc: str = "utf-8",
@@ -1598,6 +1620,7 @@ class GetMessagesByPageByGroups(SSEHandlerClassBase):
             sort: Enables message sorting within a group. It is not sorted between groups.
             response_formats: The format of the response
             keep_open: If true, keeps pulling for new message until don't have one outside the requested range.
+            streams: List of streams to search messages from.
             char_enc: Encoding for the byte stream.
             decode_error_handler: Registered decode error handler.
             cache: If True, all requested data from lw-data-provider will be saved to cache.
@@ -1619,12 +1642,14 @@ class GetMessagesByPageByGroups(SSEHandlerClassBase):
         self._page = page
         self._book_id = book_id
         self._groups = groups
+        self._streams = streams
         self._sort = sort
         self._response_formats = response_formats
         self._keep_open = keep_open
         self._max_url_length = max_url_length
 
         _check_list_or_tuple(self._groups, var_name="groups")
+        _check_list_or_tuple(self._streams, var_name="streams")
 
     def _get_urls(self, data_source: DataSource):
         page = _get_page_object(self._book_id, self._page, data_source)
@@ -1640,6 +1665,7 @@ class GetMessagesByPageByGroups(SSEHandlerClassBase):
             start_timestamp=self._start_timestamp,
             end_timestamp=self._end_timestamp,
             groups=self._groups,
+            streams=self._streams,
             response_formats=self._response_formats,
             keep_open=self._keep_open,
             sort=self._sort,