+++ release_notes.md
@@ -89,6 +89,7 @@
 3. [TH2-4924] Added GetMessagesByPages command.
 4. [TH2-4926] Added GetEventsByPages command.
 5. [TH2-4952] Added DownloadMessagesByPageByGroups, DownloadMessagesByPage and DownloadMessagesByBookByGroups commands.
+6. [Th2-4975] Added streams parameter to download and messages by groups sse commands.
 
 ## BugFixes
 1. [TH2-4925] Fix BrokenEvent and BrokenMessage.